+++ app/_config.ts
@@ -1,0 +1,8 @@
+import { TimerConfig } from "./_types";
+
+export const defaultTimerConfig: TimerConfig = {
+	pomodoroMinutes: 25,
+	shortMinutes: 5,
+	longMinutes: 15,
+	numberOfRounds: 4,
+};
