+++ app/_store.ts
@@ -1,23 +1,7 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-
-export type TimerConfig = {
-	pomodoroMinutes: number;
-	shortMinutes: number;
-	longMinutes: number;
-	numberOfRounds: number;
-	autoStartBreak: boolean;
-	autoStartWork: boolean;
-};
-
-export const defaultTimerConfig: TimerConfig = {
-	pomodoroMinutes: 25,
-	shortMinutes: 5,
-	longMinutes: 15,
-	numberOfRounds: 4,
-	autoStartBreak: false,
-	autoStartWork: false,
-};
+import { TimerConfig } from "./_types";
+import { defaultTimerConfig } from "./_config";
 
 interface TimerConfigState {
 	timerConfig: TimerConfig;
