+++ app/_types.ts
@@ -1,0 +1,6 @@
+export type TimerConfig = {
+	pomodoroMinutes: number;
+	shortMinutes: number;
+	longMinutes: number;
+	numberOfRounds: number;
+};
