+++ app/components/timer/index.module.scss
@@ -3,8 +3,10 @@
 .sectionTimer {
 	width: 100%;
 	display: flex;
+	flex-direction: column;
 	justify-content: center;
 	align-items: center;
+	gap: 1rem;
 }
 
 .backgroundTimer {
@@ -15,7 +17,7 @@
 	padding: 2.4rem;
 	display: flex;
 	flex-direction: column;
-	justify-content: space-between;
+	justify-content: space-around;
 }
 
 .typeTimes {
@@ -63,14 +65,19 @@
 	display: flex;
 	align-items: center;
 	gap: 3.2rem;
+
+	button {
+		background-color: transparent;
+		border: 0;
+		cursor: pointer;
+	}
 }
 
-.playPauseControl {
-	background-color: transparent;
-	border: 0;
+.backgroundRounds {
+	width: 64rem;
 }
 
-.restartControl {
-	background-color: transparent;
-	border: 0;
+.rounds {
+	@include text(6);
+	color: var(--primary-600);
 }
